# storage: ignore non-intent locks in `intentInterleavingIter`