Informs #100193.
Closes #109644.

This commit updates the intentInterleavingIter to ignore locks in the lock table keyspace with strengths other than lock.Intent (i.e. shared and exclusive locks). Future versions of the iterator may expose information to users about whether any non-intent locks were observed and, if so, which keys they were found on. For now, no such information is exposed.

This change is needed for replicated locks. Non-locking reads will not interact with these locks at all, so we want the intentInterleavingIter to ignore them. Meanwhile, locking reads and writes will use a different abstraction to peek into the lock table and look for conflicts.

<details>
  <summary>Benchmark delta without Locks</summary>
  
  ```
name                                                                     old time/op    new time/op    delta
IntentInterleavingIterNext/version=1/intentStride=100/keyLen=100-10         112ns ± 7%     111ns ± 4%    ~     (p=0.332 n=25+22)
IntentInterleavingIterNext/version=1/intentStride=1000000/keyLen=10-10     61.6ns ± 4%    61.1ns ± 2%    ~     (p=0.340 n=24+20)
IntentInterleavingIterNext/version=1/intentStride=1000000/keyLen=100-10    66.2ns ± 5%    66.1ns ± 4%    ~     (p=0.788 n=25+25)
IntentInterleavingIterNext/version=5/intentStride=1/keyLen=100-10           115ns ±10%     114ns ± 4%    ~     (p=0.367 n=25+25)
IntentInterleavingIterNext/version=5/intentStride=1000000/keyLen=10-10     58.9ns ± 4%    59.7ns ± 4%    ~     (p=0.051 n=21+25)
IntentInterleavingIterNext/version=5/intentStride=1000000/keyLen=100-10    63.7ns ± 3%    64.2ns ± 4%    ~     (p=0.100 n=24+25)
IntentInterleavingIterPrev/version=1/intentStride=100/keyLen=100-10         123ns ± 8%     124ns ± 4%    ~     (p=0.081 n=25+24)
IntentInterleavingIterNext/version=5/intentStride=100/keyLen=100-10         109ns ± 4%     110ns ± 4%  +0.81%  (p=0.002 n=24+24)
IntentInterleavingIterNext/version=5/intentStride=100/keyLen=10-10          100ns ± 4%     102ns ± 1%  +1.49%  (p=0.003 n=24+20)
IntentInterleavingIterPrev/version=5/intentStride=1/keyLen=10-10           84.9ns ± 4%    86.3ns ± 4%  +1.60%  (p=0.000 n=24+24)
IntentInterleavingIterPrev/version=1/intentStride=1000000/keyLen=100-10     122ns ± 6%     124ns ± 4%  +1.90%  (p=0.001 n=25+24)
IntentInterleavingIterPrev/version=5/intentStride=1000000/keyLen=100-10     123ns ± 4%     126ns ± 7%  +2.44%  (p=0.000 n=24+24)
IntentInterleavingIterNext/version=1/intentStride=100/keyLen=10-10          101ns ± 4%     104ns ± 5%  +2.46%  (p=0.000 n=25+24)
IntentInterleavingIterNext/version=5/intentStride=1/keyLen=10-10            103ns ± 6%     105ns ± 5%  +2.64%  (p=0.000 n=23+24)
IntentInterleavingIterPrev/version=5/intentStride=100/keyLen=100-10         122ns ± 3%     126ns ± 5%  +2.75%  (p=0.000 n=22+25)
IntentInterleavingIterPrev/version=5/intentStride=1/keyLen=100-10          92.0ns ± 2%    94.8ns ± 6%  +3.13%  (p=0.000 n=23+24)
IntentInterleavingIterPrev/version=1/intentStride=1000000/keyLen=10-10      112ns ± 4%     116ns ± 6%  +3.24%  (p=0.000 n=25+25)
IntentInterleavingIterPrev/version=1/intentStride=100/keyLen=10-10          112ns ± 5%     116ns ± 4%  +3.36%  (p=0.000 n=25+25)
IntentInterleavingIterPrev/version=5/intentStride=100/keyLen=10-10          112ns ± 3%     116ns ± 4%  +3.65%  (p=0.000 n=25+25)
IntentInterleavingIterNext/version=1/intentStride=1/keyLen=100-10           112ns ± 8%     117ns ± 1%  +3.83%  (p=0.000 n=25+20)
IntentInterleavingIterPrev/version=5/intentStride=1000000/keyLen=10-10      111ns ± 2%     116ns ± 3%  +4.02%  (p=0.000 n=21+24)
IntentInterleavingIterNext/version=1/intentStride=1/keyLen=10-10            103ns ± 4%     107ns ± 3%  +4.07%  (p=0.000 n=25+21)
IntentInterleavingIterPrev/version=1/intentStride=1/keyLen=100-10           105ns ± 3%     111ns ± 6%  +6.23%  (p=0.000 n=24+25)
IntentInterleavingIterPrev/version=1/intentStride=1/keyLen=10-10           94.4ns ± 3%   100.5ns ± 5%  +6.48%  (p=0.000 n=23+25)

name                                                                     old alloc/op   new alloc/op   delta
IntentInterleavingIterNext/version=1/intentStride=1/keyLen=10-10            0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=1/keyLen=100-10           0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=100/keyLen=10-10          0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=100/keyLen=100-10         0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=1000000/keyLen=10-10      0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=1000000/keyLen=100-10     0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1/keyLen=10-10            0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1/keyLen=100-10           0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=100/keyLen=10-10          0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=100/keyLen=100-10         0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1000000/keyLen=10-10      0.00B          0.00B         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1000000/keyLen=100-10     0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1/keyLen=10-10            0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1/keyLen=100-10           0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=100/keyLen=10-10          0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=100/keyLen=100-10         0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1000000/keyLen=10-10      0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1000000/keyLen=100-10     0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1/keyLen=10-10            0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1/keyLen=100-10           0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=100/keyLen=10-10          0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=100/keyLen=100-10         0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1000000/keyLen=10-10      0.00B          0.00B         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1000000/keyLen=100-10     0.00B          0.00B         ~     (all equal)

name                                                                     old allocs/op  new allocs/op  delta
IntentInterleavingIterNext/version=1/intentStride=1/keyLen=10-10             0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=1/keyLen=100-10            0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=100/keyLen=10-10           0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=100/keyLen=100-10          0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=1000000/keyLen=10-10       0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=1/intentStride=1000000/keyLen=100-10      0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1/keyLen=10-10             0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1/keyLen=100-10            0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=100/keyLen=10-10           0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=100/keyLen=100-10          0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1000000/keyLen=10-10       0.00           0.00         ~     (all equal)
IntentInterleavingIterNext/version=5/intentStride=1000000/keyLen=100-10      0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1/keyLen=10-10             0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1/keyLen=100-10            0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=100/keyLen=10-10           0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=100/keyLen=100-10          0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1000000/keyLen=10-10       0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=1/intentStride=1000000/keyLen=100-10      0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1/keyLen=10-10             0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1/keyLen=100-10            0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=100/keyLen=10-10           0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=100/keyLen=100-10          0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1000000/keyLen=10-10       0.00           0.00         ~     (all equal)
IntentInterleavingIterPrev/version=5/intentStride=1000000/keyLen=100-10      0.00           0.00         ~     (all equal)
  ```
</details>

<details>
  <summary>Benchmarks with Locks</summary>
  
  ```
name                                                                                        time/op
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1/keyLen=10-10                149ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1/keyLen=100-10               161ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=100/keyLen=10-10              105ns ± 0%
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=100/keyLen=100-10             113ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1000000/keyLen=10-10          104ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1000000/keyLen=100-10         112ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1/keyLen=10-10              173ns ± 2%
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1/keyLen=100-10             187ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=100/keyLen=10-10           94.3ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=100/keyLen=100-10           102ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1000000/keyLen=10-10       93.0ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1000000/keyLen=100-10       100ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1/keyLen=10-10          172ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1/keyLen=100-10         188ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=100/keyLen=10-10       93.6ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=100/keyLen=100-10       101ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1000000/keyLen=10-10   54.1ns ± 1%
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1000000/keyLen=100-10  58.8ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1/lockStride=1/keyLen=10-10                112ns ± 0%
IntentInterleavingIterNext/version=5/intentStride=1/lockStride=1/keyLen=100-10               121ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1/lockStride=100/keyLen=10-10             96.6ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1/lockStride=100/keyLen=100-10             105ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1/lockStride=1000000/keyLen=10-10         96.6ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1/lockStride=1000000/keyLen=100-10         104ns ± 0%
IntentInterleavingIterNext/version=5/intentStride=100/lockStride=1/keyLen=10-10              108ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=100/lockStride=1/keyLen=100-10             118ns ± 0%
IntentInterleavingIterNext/version=5/intentStride=100/lockStride=100/keyLen=10-10           91.5ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=100/lockStride=100/keyLen=100-10          98.6ns ± 0%
IntentInterleavingIterNext/version=5/intentStride=100/lockStride=1000000/keyLen=10-10       91.2ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=100/lockStride=1000000/keyLen=100-10      98.5ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1000000/lockStride=1/keyLen=10-10          108ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1000000/lockStride=1/keyLen=100-10         118ns ± 0%
IntentInterleavingIterNext/version=5/intentStride=1000000/lockStride=100/keyLen=10-10       91.6ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1000000/lockStride=100/keyLen=100-10      98.9ns ± 1%
IntentInterleavingIterNext/version=5/intentStride=1000000/lockStride=1000000/keyLen=10-10   51.6ns ± 0%
IntentInterleavingIterNext/version=5/intentStride=1000000/lockStride=1000000/keyLen=100-10  56.5ns ± 2%
IntentInterleavingIterPrev/version=1/intentStride=1/lockStride=1/keyLen=10-10                149ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=1/lockStride=1/keyLen=100-10               159ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=1/lockStride=100/keyLen=10-10             96.7ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=1/lockStride=100/keyLen=100-10             105ns ± 5%
IntentInterleavingIterPrev/version=1/intentStride=1/lockStride=1000000/keyLen=10-10         94.2ns ± 0%
IntentInterleavingIterPrev/version=1/intentStride=1/lockStride=1000000/keyLen=100-10         103ns ± 0%
IntentInterleavingIterPrev/version=1/intentStride=100/lockStride=1/keyLen=10-10              190ns ± 0%
IntentInterleavingIterPrev/version=1/intentStride=100/lockStride=1/keyLen=100-10             207ns ± 0%
IntentInterleavingIterPrev/version=1/intentStride=100/lockStride=100/keyLen=10-10            103ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=100/lockStride=100/keyLen=100-10           111ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=100/lockStride=1000000/keyLen=10-10        102ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=100/lockStride=1000000/keyLen=100-10       110ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=1000000/lockStride=1/keyLen=10-10          190ns ± 0%
IntentInterleavingIterPrev/version=1/intentStride=1000000/lockStride=1/keyLen=100-10         207ns ± 0%
IntentInterleavingIterPrev/version=1/intentStride=1000000/lockStride=100/keyLen=10-10        102ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=1000000/lockStride=100/keyLen=100-10       111ns ± 1%
IntentInterleavingIterPrev/version=1/intentStride=1000000/lockStride=1000000/keyLen=10-10    101ns ± 0%
IntentInterleavingIterPrev/version=1/intentStride=1000000/lockStride=1000000/keyLen=100-10   110ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=1/lockStride=1/keyLen=10-10               94.7ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=1/lockStride=1/keyLen=100-10               103ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=1/lockStride=100/keyLen=10-10             77.8ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=1/lockStride=100/keyLen=100-10            86.4ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=1/lockStride=1000000/keyLen=10-10         78.5ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=1/lockStride=1000000/keyLen=100-10        86.3ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=100/lockStride=1/keyLen=10-10              123ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=100/lockStride=1/keyLen=100-10             135ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=100/lockStride=100/keyLen=10-10            103ns ± 2%
IntentInterleavingIterPrev/version=5/intentStride=100/lockStride=100/keyLen=100-10           111ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=100/lockStride=1000000/keyLen=10-10        102ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=100/lockStride=1000000/keyLen=100-10       111ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=1000000/lockStride=1/keyLen=10-10          121ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=1000000/lockStride=1/keyLen=100-10         133ns ± 2%
IntentInterleavingIterPrev/version=5/intentStride=1000000/lockStride=100/keyLen=10-10        102ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=1000000/lockStride=100/keyLen=100-10       112ns ± 1%
IntentInterleavingIterPrev/version=5/intentStride=1000000/lockStride=1000000/keyLen=10-10    102ns ± 0%
IntentInterleavingIterPrev/version=5/intentStride=1000000/lockStride=1000000/keyLen=100-10   111ns ± 0%

name                                                                                        alloc/op
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1/keyLen=10-10                0.00B
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1/keyLen=100-10               0.00B
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=100/keyLen=10-10              0.00B
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=100/keyLen=100-10             0.00B
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1000000/keyLen=10-10          0.00B
IntentInterleavingIterNext/version=1/intentStride=1/lockStride=1000000/keyLen=100-10         0.00B
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1/keyLen=10-10              0.00B
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1/keyLen=100-10             0.00B
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=100/keyLen=10-10            0.00B
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=100/keyLen=100-10           0.00B
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1000000/keyLen=10-10        0.00B
IntentInterleavingIterNext/version=1/intentStride=100/lockStride=1000000/keyLen=100-10       0.00B
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1/keyLen=10-10          0.00B
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1/keyLen=100-10         0.00B
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=100/keyLen=10-10        0.00B
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=100/keyLen=100-10       0.00B
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1000000/keyLen=10-10    0.00B
IntentInterleavingIterNext/version=1/intentStride=1000000/lockStride=1000000/keyLen=100-10   0.00B
IntentInterleavingIterNext/version=5/intentStride=1/lockStride=1/keyLen=10-10                0.00B
…